Remove svg bar graphs, replace with css.
NOTE: This is only for graphs in a flexigrid, as they
did not have the arrow, which will be more problematic
if not impossible to do with css (though maybe with js).

Load average for host is based on 5 (apparently this is
some agreed-upon temporary scale).  There is no real load
for vm pools in the database, so these are stubbed out
for now.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>