Added support for booting a VM from an ISO image.
Also added a few helper methods to Vm to contain the knowledge
of how Cobbler integration is contained.

When a user adds an ISO image to the Cobbler server on the appliance,
they will need to do so using the full NFS path for where the virtual
image will go to mount it; i.e., hostname:/path/to/filename.iso

If the filename ends in ".iso" then the virtual machine will mount the
file as a CDROM device and boot it. Otherwise, it mounts it as a hard
disk device.

To add an image to Cobbler, do the following:

1. Download an ISO image, such as the KDE LiveImage from Fedora.
2. Copy it to the NFS directory on the server:
  cp *.iso /ovirtnfs/kde-live-cd.iso
3. Add that image to your Cobbler instance:
  cobbler image add --name=KDE-LiveCD --file=management.priv.ovirt.org:/ovirtnfs/kde-live-cd.iso
4. Create a new VM in your server.
5. Select "KDE-LiveCD" from the list of operating systems.
6. Save the VM.
7. Start the VM.

It should run the selected ISO.