Call setup provisioning method from create as well as edit and pass mac as arg
Addition of the edit vm screen broke provisioning of vms on create since the
format of the provisioning names changed.  The names were no longer being
prefixed with the words cobbler or profile.  This caused taskomatic to
skip the section where the cobbler system create was done meaning the vm
would never be provisioned.

Fix is to have create vm call the provisioning method the same way edit
does.  The taskomatic stuff can probably be removed for the time being,
but I'm going to leave it in since I'm not sure this is the right way
to solve this problem long term.

Also, the setup provisioning method did not associate the MAC of the
new system to the cobbler system making that call useless.  This is
also fixed now.

This fix is being committed without review since it fixes a major bug
and we're wrapping up for a release.

Signed-off-by: Perry Myers <pmyers@redhat.com>