Fixed the tests for host-browser.
A change to the identify protocol was overlooked. When an empty value was
submitted, host-browser would throw an exception. But, a change allowed
empty values to get by without notice. Fixed.

Also added a Requires dependency on ruby-flexmock so that the RPM is
correctly installed when building the server RPM.

Finally, reintroduced a line that was deleted by mistaken from NicControllerTest.