Replaced the config scripts with configuration encoding.
Rather than sending the node a series of scripts that load
kernel modules, or are tightly coupled to tools like augeas,
this patch introduces an encoding scheme for data.

A line that begins with "bonding" describes a bonded interface.
From it the configuration processor generates a simple bonding
file that includes an alias for the bonding kernel module

A line that begins with "ifcfg" describes an network
interface. It will contain the mac address and interface name,
followed by all needed configuration values to bring the
interface up.