initial model work for lvm storage (revised2)
This includes the model classes w/ migrations for the lvm storage pools and volumes, and some changes required to support Storage Volume tasks. This patch does not include all of the necessary model API methods to support lvm, but it's a starting point -- some changes to clalance's taskomatic bits will be required to support this, and we will need additional model enhancements when the UI work is done, and when the taskomatic back end is finalized.

revised to fix a couple controller bugs and to increment the migration version #
and to fix some additional model problems in taskomatic and host-status