Version 1 of Revamped Tree Navigation. Take 3
The new javascript tree widget contains the following features/changes from previous implementation:

* The html for the list is dynamically generated using a javascript template system.  This will allow us to plug in different layouts per tree as the widget matures.
* Updates to the tree are now incremental, rather than a full rip and replace as earlier. We have 2 states we currently look for - 'new' and 'changed'.  The first generates new html and appends it to the DOM, the second just does a replacement of the content of existing nodes.
* Vastly simplified the markup and css.
* Added calls where appropriate to refresh the tree before next planned call (for instance, if you add a new hardware pool).
* Added slide effect when opening and closing a node of the tree.
* Clicking the plus/minus opens/close the node only, does not load main content area.
* Clicking anywhere to the right of that on a given node will load content area.
* Added interim icons for 'smartpool' and 'add smartpool'

Note that aside from the nav area, this should not impact the existing trees which have not been converted yet (all popups that have one), as this is a completely separate codebase with it's own js and css files.

Related, but not technically part of the tree, I added a choose_layout method to allow testing of javascript components as we are building them to help eliminate possible side effects from other code.  When not in a production environment, you can pass in ?component_layout=[name] where [name] is the name of a shell rhtml file you have put in views/layouts/components.  As our UI is growing increasingly complex, I think this will be a very useful way to facilitate building components.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>