move dialog fix.
fixes a problem in the move (and add pool to smart pool) dialogs when permission was granted to the top level DirectoryPool rather than to individual HW pools. This code will be going away soon anyway, as we need to migrate the dialog tree view away from the legacy implementation to jguiditta's new tree code.

This fixes BZ 464273 and 464398.