Fix bug in host-status.
This patch fixes a bug in host-status where the connection was being closed
at the end of the loop instead of after so consequent calls were failing
making it think the vm was off.

There's also a little extra debugging prints here and another small fix
to return if we can't get a domain list.

Signed-off-by: Ian Main <imain@redhat.com>