Test/fixture revamp.
Mo, thanks for testing, I have rebased and merged the patch with Scott's
new stuff, should still be fine.  Please give it a go, and I will push
when I hear back from you.

Redoing the test fixtures to take advantage of rails 2.1 'foxy fixtures'.
This means very few ids, except for a couple special cases.  This not only
gets us closer to current rails standards, but makes the fixtures more readable
and maintainable, which hopefully translates into making it less painful to
write comprehensive tests for new code (and build out better tests for
existing code).

A side effect of this is that I have been trimming unneeded fixture entries
and renaming many of the ones we are keeping so it is more clear what they are,
as well as general cleanup of incorrect things like quotes where they shouldn't
be.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>