Replace host-status with db-omatic.
This patch replaces host-status with db-omatic (formerly qpid-db-sync).
This uses the qpid console to recieve events from libvirt-qpid on various
nodes and update state in the database directly.  It also uses heartbeating
from the libvirt-qpid agents so it knows when a host or agent dies.

This update fixes a bug that occurs when a domain is user created; I was
using 'next' instead of 'return'.

Signed-off-by: Ian Main <imain@redhat.com>