This patch implements NFS file creation/deletion in the taskomatic back-end.  Actually, this patch does quite a bit more than that:
1)  Implement NFS file creation/deletion
2)  Fix LVM creation - due to the way I was testing, there was a glaring
bug in the implementation that caused it not to work at all on new, raw LUNs
3)  Make sure to flip the Pool and Volume states to "available", once
things are available.

With this patch in place, along with a minor patch to libvirt
(updated packages coming soon), I was able to successfully create
and delete LVM volumes, NFS files, and then use those newly created
LUNs/files in a guest.

Caveat: note that NFS file creation can take a *long* time.  Basically,
we fully allocate the file at creation time, which means we have to
write the entire disk full of zeros.  During this time, both
libvirtd (on the remote node) and taskomatic can't accept any more tasks.
This will eventually be fixed by a) making libvirtd multithreaded, and
b) making taskomatic fully threaded.

Finally: with this in place, a couple of problems come to light.
One is that we probably want to give users the ability to choose a sparse
file vs. a non-sparse file (but that's a minor change in the frontend and
in taskomatic).  The second is much more complicated and important;
namely, that we need users to be able to assign /dev/sda to iSCSI LUN 3,
/dev/sdb to NFS file 2, etc.  At the *very* least we need users to be
able to specify which device is the root device.  In any case, this is
future work.

Signed-off-by: Chris Lalancette <clalance@redhat.com>