additional changes for  Bug 466719.
Fixed incorrect column name for load_average -- so sorting works now. Also removed the ability to sort on num_cpus and cpu_speed. These are now stored in a separate object, so the prior code for this was broken, and it doesn't seem like it would be a high priority to support sorting on these fields.