Fix smart pools in nav.
They were not updating when you added a new one because We were
only recursing down through children if an element was not changed,
rather than always if there were children. If I recall correctly,
this was originally to save processing time, but with recent
changes, it is no longer accurate.

Also realized that tree_controller was not splitting properly based
on name, so if you had a name with '-'s in it, it would always return
as 'changed'. In addition, the 'smart root' never changes (at least by
the user) so make state='unchanged'.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>