Add LVM volume scanning back into taskomatic.
This patch adds LVM volume scanning back into taskomatic with the help
of some libvirt-qpid changes that set the LVM volume group name as a
property of the physical volume if applicable.

Also included is a small fix to qmf-libvirt-example to properly sort
volumes by storage pool.  This script has been very helpful in debugging
qpid issues both for developers and users.

Signed-off-by: Ian Main <imain@redhat.com>
Acked-by: Chris Lalancette <clalance@redhat.com>