Persistantly store vm / host history
This patch adds a table and model classes to persistently
store the history of which vms were running on which hosts.
Changes made to dbomatic to record history on changes to
a vm's state. Fields that are recorded include vm / host ids,
time started and ended, vm state, and vnc port of the
vm when running. Additional fields will most likely
be desired, and can easily be added here or in subsequent
patches.