Fix VM shutdown
This patch fixes VM shutdown.  Previously when shutting down a vm we
just set it to 'stopped'.  However, this is not always the case as it's
possible for the guest OS to ignore the shutdown event.  Also it is
possible for a VM to be stopped by the user in which case it would be
nice to set it stopped in the DB and clear the flags.

Signed-off-by: Ian Main <imain@redhat.com>