Fixes to the managed_node_configuration code.
Added public APIs to Bonding and Nic to retrieve networking details for
the respective interface.

Refactored how interfaces, bonded interfaces and bridges are defined.

When returning a configuration for the node, a bridge is first created,
using the networking definition for the device. The network interface is
then bridged over that device.

When a bonded interface is created, a bridge is created for it as well.

If a hostname is unknown, or it does not send up any macs, then it is
returned an empty configuration payload and not an error status code.