First step of detail pane UI cleanup.
This patch implements layout fixes to clean up look of the
details pane throughout the site, so it looks more like
the design here:
http://ovirt.org/page/Summary_UX

Note that this does not include the actual rearrangement of
Hardware Pools summary pages.  Best place to see changes is
on the summary for a Resource Pool.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>