Revised attempt at refactoring the vm_controller using the service layer. (second revision)
This code is based on David Lutterkort's proposal outlined in
https://www.redhat.com/archives/ovirt-devel/2009-April/msg00086.html

I've included service-layer methods for show, create, update, destroy, vm_action, and cancel_queued_tasks. These are the concrete actions that make sense as part of the API. Most of the remaining vm_controller methods are pure-WUI actions -- for things such as generating the web form for a new VM (the subsequent 'create' action _is_ part of the API).

There's a certain amount of temporary code that we need in place while only some of the controllers have been converted, as we still need permissions pre-filters for non-converted create, etc. methods -- but for the converted controllers we don't want to perform actions in before_filters that will be done as part of the services layer. Any temporary methods, etc. that will go away when the last controller is converted have been marked clearly with FIXME comments.

Revised version passes perm_obj arg into set_perms -- and always uses this method to set @perm_obj rather than setting it directly. I've also cleaned the permission-setting calls in general, removing more of the no-longer-valid redirects, etc.

Also revised to remove a bunch of unnecessary inline calls to set @current_pool_id and removed an unnecessary FIXME comment.