Cloud UI V1 (readonly).
This patch creates the shell for much of what the cloud ui will
become.  It is focused on layout (lightly fedora themed), accessibility,
and 'bookmarkability'.  This layout uses no javascript (that will be
used to enhance interactivity in a future patch).  It also introduces
the idea of using a simple table rather than the more conplex js/json
we use in the admin side.  Plan here is to gradually enhance this as
needed to make more interactive.

CSS was kept separate from admin CSS to facilitate theming by
users/deployments.

Controllers are designed (even though large parts are stubbed) to be
'skinny' and leverage as much as possible the new service api.  There
are a few readonly AR calls that in near future could easily be moved to
service layer to be even simpler calls.  They are kept separate from admin
things so they can be as simple as possible as well a making it conceivable
that one could deploy the cloud part of the app without also deplying admin
(perhaps one wants to deply that elsewhere).  For now this would have to be
done manually, but perhaps in future we can setup the service layer/ar classes
to be a separate rpm that two different wuis (admin/cloud) could call into,
enhancing modularity.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>