Starting of new ovirt QMF API.
This is the new ovirt agent that implements the QMF API.  This is just
its humble beginnings.  The goal was to implement the create VM method on
VMs but it turns out we need to specify the VM pool in order to do that.
This is currently hardcoded to '5' and works locally but would likely not
work elsewhere.  What does work well is listing/querying of existing VMs.

There is still lots of work to do here, this is just so that we have a
common place to work on going forward.  One of the first things to do
is to split out the Ovirt and VmDef classes into seperate files and
implement VM pools.

Signed-off-by: Ian Main <imain@redhat.com>