Use fixed mount points and add timeouts to various calls.
This patch uses fixed mount points necessary for migration to work
properly.  Mount points are unique for each storage type.

This also uses the new :timeout keyword argument for various operations
that could take a while.  This should fix the 'seq' timeout problem
we've been seeing.  This requires the latest ruby-qpid which is now
in the ovirt repo.

Signed-off-by: Ian Main <imain@redhat.com>