Set up ovirt-agent so it starts as a daemon
This sets up ovirt-agent to start as a daemon along with dbomatic,
taskomatic etc.  While this is not ready for prime time I think it
should be ok as you still require a valid, authenticated, connection
to qpidd to access it.

This update adds the conf/ovirt-agent start/stop script.

Signed-off-by: Ian Main <imain@redhat.com>