Update core app config for Rails 2.3.2
This patch updates app config and makes any
required filename changes to support rails 2.3.2.

Notable changes are:
* application.rb -> application_controller.rb
* cleanup of environment.rb
* new config/initializers dir (this is where much of
  the environment.rb stuff went)
* gettext/rails -> gettext_rails

Also note that applying this patch makes ovirt server
no longer run on Fedora 10, as Rails 2.3.2 is not in that
yum repo.  If you wish to run the app on F10 after this
upgrade, you can probably still do so with:
gem update rails
but of course this has the potential to cause issues with
yum and gem conflicting.

Signed-off-by: Jason Guiditta <jason.guiditt@gmail.com>
Signed-off-by: Jason Guiditta <jason.guiditta@gmail.com>