Use volume key instead of path to identify volume.
This patch teaches taskomatic to use the volume 'key' instead of the
path from libvirt to key the volume off of in the database.  This fixes
the duplicate iscsi volume bug we were seeing.  The issue was that
libvirt changed the way they name storage volumes and included a local
ID that changed each time it was attached.

This patch now adds a 'key' field to the storage_volumes table and
uses that to key off for various taskomatic related activities.

Note that the first run with this new patch will cause duplicate
volumes because of the key change.  Also because volumes are now
looked up by 'key' any VMs using old volumes will not work, so you
will have to delete all your VMs and rescan your storage volumes to
make ovirt work properly again.

Signed-off-by: Ian Main <imain@redhat.com>