Wrapper for mongrel to run with rails 2.3.2
This file is needed because we use --prefix when running
mongrel_rails (which I think we need for now).  It wraps
the AbstractRequest call that is no longer in rails (but
triggered by --prefix) so mongrel does not constantly die.
It is possible there is a way now in rails itself to specify
something in place of this flag, but I have not been able
to find where to put that setting yet.

Signed-off-by: Jason Guiditta <jason.guiditta@gmail.com>