hack betternested set to enable pool deletion
This is an ugly hack to fix the betternestedset plugin as currently
a bug does not allow pools to be deleted. The single line commented
out in this patch deletes the pool itself instead of only its children
for whatever reason.

This wouldn't be a problem were it not for the recent
change to active record which enforces optimistic locking upon
destroy operations. Since the database record is mistakingly
deleted in betternestedset, via the 'before_destroy' hook,
the record does not exist when the activerecord destroy operation is actually
performed and thus an exception is thrown at

  /usr/lib/ruby/gems/1.8/gems/activerecord-2.3.2/lib/active_record/locking/optimistic.rb
    line 100

Note this patch only works since we explicitly disable a hw pool from being deleted
unless all its children are in the oVirt hardware_pool_server. It is ugly but works until
something more elegant can be found.